@@ -9,8 +9,9 @@ environment:
   sdk: ">=2.14.4 <3.0.0"
 
 dependencies:
-  awesome_notifications: ^0.7.4+1
-  awesome_notifications_fcm: ^0.7.4+1
+  awesome_notifications_core: ^0.8.1
+  awesome_notifications: any
+  awesome_notifications_fcm: any
   badges: ^2.0.2
   barcode_widget: ^2.0.3
   barcode_image: ^2.0.2
@@ -28,7 +29,7 @@ dependencies:
   dio_cache_interceptor: ^3.2.6
   dio_cache_interceptor_hive_store: ^3.1.1
   dotted_border: ^2.0.0+2
-  extended_image: ^6.0.2+1
+  extended_image: ^8.1.1
   firebase_core: 2.11.0
   firebase_analytics: 10.2.1
   firebase_in_app_messaging: ^0.7.0+10
@@ -54,16 +55,16 @@ dependencies:
   in_app_review: ^2.0.3
   logger: ^1.3.0
   mdi: ^5.0.0-nullsafety.0
-  package_info_plus: ^3.0.2
+  package_info_plus: ^4.2.0
   path_provider: ^2.0.2
   permission_handler: ^10.2.0
   photo_view: ^0.14.0
   qr_flutter: ^4.0.0
   recase: ^4.0.0
   responsive_framework: ^0.2.0
-  screenshot: ^1.2.3
-  sentry_flutter: ^6.18.2
-  share_plus: ^6.3.0
+  screenshot: ^2.1.0
+  sentry_flutter: ^7.10.1
+  share_plus: ^7.2.1
   simple_gesture_detector: ^0.2.0
   url_launcher: ^6.1.7
   video_player: ^2.6.1
