@@ -1,20 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:get/route_manager.dart';
 import 'package:mi_utem/config/routes.dart';
-import 'package:mi_utem/screens/asignatura_detalle_screen.dart';
-import 'package:mi_utem/screens/asignaturas_lista_screen.dart';
 import 'package:mi_utem/controllers/asignaturas_controller.dart';
 import 'package:mi_utem/controllers/qr_passes_controller.dart';
+import 'package:mi_utem/controllers/user_controller.dart';
+import 'package:mi_utem/screens/asignatura_detalle_screen.dart';
+import 'package:mi_utem/screens/asignaturas_lista_screen.dart';
 import 'package:mi_utem/screens/calculadora_notas_screen.dart';
 import 'package:mi_utem/screens/credencial_screen.dart';
 import 'package:mi_utem/screens/horario/horario_screen.dart';
 import 'package:mi_utem/screens/login_screen/login_screen.dart';
-import 'package:mi_utem/screens/main_screen.dart';
+import 'package:mi_utem/screens/main/main_screen.dart';
 import 'package:mi_utem/screens/permiso_covid_screen.dart';
 import 'package:mi_utem/screens/splash_screen.dart';
 import 'package:mi_utem/screens/usuario_screen.dart';
-import 'package:mi_utem/services/auth_service.dart';
-import 'package:mi_utem/services/perfil_service.dart';
 import 'package:mi_utem/widgets/acerca_screen.dart';
 
 final _loginPage = GetPage(
@@ -26,11 +25,7 @@ final _loginPage = GetPage(
 final _homePage = GetPage(
   name: Routes.home,
   bindings: [QrPassesBinding()],
-  page: () {
-    final usuario = PerfilService.getLocalUsuario();
-
-    return MainScreen(usuario: usuario);
-  },
+  page: () => MainScreen(),
   middlewares: [OnlyAuthMiddleware()],
 );
 
@@ -86,7 +81,7 @@ final pages = <GetPage>[
 class OnlyAuthMiddleware extends GetMiddleware {
   @override
   RouteSettings? redirect(String? page) {
-    final isLoggedIn = AuthService.isLoggedIn();
+    final isLoggedIn = UserController.to.isLoggedIn;
     if (!isLoggedIn) {
       return const RouteSettings(name: Routes.login);
     }
@@ -97,7 +92,7 @@ class OnlyAuthMiddleware extends GetMiddleware {
 class OnlyNoAuthMiddleware extends GetMiddleware {
   @override
   RouteSettings? redirect(String? page) {
-    final isLoggedIn = AuthService.isLoggedIn();
+    final isLoggedIn = UserController.to.isLoggedIn;
     if (isLoggedIn) {
       return const RouteSettings(name: Routes.home);
     }
