@@ -11,17 +11,15 @@ import 'package:get_storage/get_storage.dart';
 import 'package:mi_utem/config/constants.dart';
 import 'package:mi_utem/config/router.dart';
 import 'package:mi_utem/config/routes.dart';
-import 'package:mi_utem/controllers/calculator_controller.dart';
 import 'package:mi_utem/controllers/carreras_controller.dart';
-import 'package:mi_utem/services/analytics_service.dart';
-import 'package:mi_utem/services/auth_service.dart';
+import 'package:mi_utem/controllers/user_controller.dart';
 import 'package:mi_utem/services/background_service.dart';
-import 'package:mi_utem/services/config_service.dart';
 import 'package:mi_utem/services/notification_service.dart';
-import 'package:mi_utem/services/perfil_service.dart';
 import 'package:mi_utem/themes/theme.dart';
 import 'package:responsive_framework/responsive_framework.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
+
+import 'services/remote_config/remote_config.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -30,7 +28,7 @@ void main() async {
   await GetStorage.init();
   await SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]);
   await Firebase.initializeApp();
-  await ConfigService.getInstance();
+  await RemoteConfigService.initialize();
   await NotificationService.initialize();
   await BackgroundService.initAndStart();
   await SentryFlutter.init(
@@ -43,26 +41,8 @@ void main() async {
   );
 }
 
-class MiUtem extends StatefulWidget {
-  @override
-  State<MiUtem> createState() => _MiUtemState();
-}
-
-class _MiUtemState extends State<MiUtem> {
+class MiUtem extends StatelessWidget {
   final FirebaseAnalytics analytics = FirebaseAnalytics.instance;
-
-  final calculatorController = Get.put(CalculatorController());
-
-  @override
-  void initState() {
-    if (AuthService.isLoggedIn()) {
-      final user = PerfilService.getLocalUsuario();
-      AnalyticsService.setUser(user);
-    } else {
-      AnalyticsService.removeUser();
-    }
-    super.initState();
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -80,6 +60,7 @@ class _MiUtemState extends State<MiUtem> {
       debugShowCheckedModeBanner: false,
       title: 'Mi UTEM',
       initialBinding: BindingsBuilder(() {
+        Get.put(UserController(), permanent: true);
         Get.put(CarrerasController(), permanent: true);
       }),
       theme: MainTheme.theme,
