@@ -2,7 +2,6 @@ import 'dart:convert';
 import 'dart:math';
 
 import 'package:dio/dio.dart';
-import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
@@ -10,11 +9,11 @@ import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:get/get.dart';
 import 'package:mi_utem/config/routes.dart';
+import 'package:mi_utem/controllers/user_controller.dart';
 import 'package:mi_utem/helpers/snackbars.dart';
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/services/analytics_service.dart';
-import 'package:mi_utem/services/auth_service.dart';
-import 'package:mi_utem/services/config_service.dart';
+import 'package:mi_utem/services/remote_config/remote_config.dart';
 import 'package:mi_utem/widgets/acerca_dialog.dart';
 import 'package:mi_utem/widgets/dialogs/monkey_error_dialog.dart';
 import 'package:mi_utem/widgets/dialogs/not_ready_dialog.dart';
@@ -37,12 +36,9 @@ class _LoginScreenState extends State<LoginScreen> {
 
   GlobalKey<FormState> _formKey = GlobalKey<FormState>();
 
-  FirebaseRemoteConfig? _remoteConfig;
-
   @override
   void initState() {
     super.initState();
-    _remoteConfig = ConfigService.config;
 
     SystemChrome.setSystemUIOverlayStyle(
       SystemUiOverlayStyle(
@@ -99,9 +95,7 @@ class _LoginScreenState extends State<LoginScreen> {
   }
 
   String get _creditText {
-    List<dynamic> texts = jsonDecode(_remoteConfig!.getString(
-      ConfigService.CREDITOS,
-    ));
+    List<dynamic> texts = jsonDecode(RemoteConfigService.creditos);
 
     Random random = new Random();
 
@@ -251,8 +245,8 @@ class _LoginScreenState extends State<LoginScreen> {
         );
 
         try {
-          bool esPrimeraVez = await AuthService.esPrimeraVez();
-          Usuario usuario = await AuthService.login(correo, contrasenia, true);
+          bool isFirstTime = UserController.to.isFirstTime;
+          Usuario usuario = await UserController.to.login(correo, contrasenia);
 
           AnalyticsService.logEvent('login');
           AnalyticsService.setUser(usuario);
@@ -261,7 +255,7 @@ class _LoginScreenState extends State<LoginScreen> {
             Routes.home,
           );
 
-          if (esPrimeraVez) {
+          if (isFirstTime) {
             Get.dialog(
               AcercaDialog(),
             );
