@@ -1,18 +1,13 @@
-import "dart:convert";
-import "dart:math";
-
-import "package:firebase_remote_config/firebase_remote_config.dart";
 import 'package:flutter/foundation.dart';
 import "package:flutter/material.dart";
 import "package:flutter/services.dart";
-import "package:flutter_markdown/flutter_markdown.dart";
-import "package:get/get.dart";
 import 'package:mi_utem/controllers/grades_changes_controller.dart';
-import "package:mi_utem/models/usuario.dart";
 import "package:mi_utem/screens/beca_alimentacion/cupon.dart";
-import "package:mi_utem/services/config_service.dart";
+import "package:mi_utem/screens/main/widgets/greetings.dart";
 import "package:mi_utem/services/perfil_service.dart";
+import "package:mi_utem/services/remote_config/remote_config.dart";
 import "package:mi_utem/services/review_service.dart";
+import 'package:mi_utem/widgets/banners_section.dart';
 import "package:mi_utem/widgets/custom_app_bar.dart";
 import "package:mi_utem/widgets/custom_drawer.dart";
 import "package:mi_utem/widgets/noticias_carrusel.dart";
@@ -20,20 +15,16 @@ import "package:mi_utem/widgets/permisos_section.dart";
 import "package:mi_utem/widgets/quick_menu_section.dart";
 
 class MainScreen extends StatefulWidget {
-  final Usuario usuario;
-  MainScreen({Key? key, required this.usuario}) : super(key: key);
+  MainScreen({Key? key}) : super(key: key);
 
   @override
   _MainScreenState createState() => _MainScreenState();
 }
 
 class _MainScreenState extends State<MainScreen> {
-  late FirebaseRemoteConfig? _remoteConfig;
-
   @override
   void initState() {
     super.initState();
-    _remoteConfig = ConfigService.config;
     PerfilService.saveFcmToken();
     SystemChrome.setSystemUIOverlayStyle(
       SystemUiOverlayStyle(
@@ -54,22 +45,13 @@ class _MainScreenState extends State<MainScreen> {
     super.dispose();
   }
 
-  String get _greetingText {
-    List<dynamic> texts = jsonDecode(_remoteConfig!.getString(
-      ConfigService.GREETINGS,
-    ));
-
-    Random random = new Random();
-    String text = texts[random.nextInt(texts.length)];
-    text = text.replaceAll("%name", widget.usuario.primerNombre);
-    return text;
-  }
-
   @override
   Widget build(BuildContext context) {
+    final banners = RemoteConfigService.banners;
+
     return Scaffold(
       appBar: CustomAppBar(title: Text("Inicio")),
-      drawer: CustomDrawer(usuario: widget.usuario),
+      drawer: CustomDrawer(),
       backgroundColor: Colors.grey[200],
       floatingActionButton: kDebugMode
           ? FloatingActionButton(
@@ -92,22 +74,18 @@ class _MainScreenState extends State<MainScreen> {
               //code: null,
               code: "3N14M3ARCX",
             ),
-            Container(
-              padding: EdgeInsets.symmetric(horizontal: 20),
-              width: double.infinity,
-              child: MarkdownBody(
-                data: _greetingText,
-                styleSheet: MarkdownStyleSheet(
-                  p: Get.textTheme.displayMedium!
-                      .copyWith(fontWeight: FontWeight.normal),
-                ),
-              ),
-            ),
+            MainScreenGreetings(),
             Container(height: 20),
             PermisosCovidSection(),
             Container(height: 20),
             QuickMenuSection(),
             Container(height: 20),
+            if (banners.isNotEmpty) ...[
+              BannersSection(
+                banners: banners,
+              ),
+              Container(height: 20),
+            ],
             NoticiasSection(),
           ],
         ),
