@@ -2,8 +2,8 @@ import 'package:dio/dio.dart';
 import 'package:dio_http_cache/dio_http_cache.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:mi_utem/controllers/grades_changes_controller.dart';
+import 'package:mi_utem/controllers/user_controller.dart';
 import 'package:mi_utem/models/grades.dart';
-import 'package:mi_utem/services/perfil_service.dart';
 import 'package:mi_utem/utils/dio_miutem_client.dart';
 
 class GradesService {
@@ -17,7 +17,7 @@ class GradesService {
     bool saveGrades = true,
   }) async {
     final uri = "/v1/carreras/$carreraId/asignaturas/$asignaturaId/notas";
-    final user = PerfilService.getLocalUsuario();
+    final user = UserController.to.user.value;
 
     Response response = await _dio.get(
       uri,
@@ -25,7 +25,7 @@ class GradesService {
         Duration(days: 1),
         maxStale: Duration(days: 7),
         forceRefresh: forceRefresh,
-        subKey: user.rut?.numero.toString(),
+        subKey: user?.rut?.numero.toString(),
       ),
     );
 
